Remove currently unused data provider, validate file during metadata process
