+++ lib/galaxy/datatypes/dataproviders/column.py
@@ -4,10 +4,8 @@ is further subdivided into multiple data (e.g. columns from a line).
 """
 import logging
 import re
-from typing import Dict
 from urllib.parse import unquote_plus
 
-from galaxy.datatypes.goldenpath import AGPFile
 from . import line
 
 _TODO = """
@@ -355,19 +353,3 @@ class DictDataProvider(ColumnarDataProvider):
         for column_values in parent_gen:
             map = dict(zip(self.column_names, column_values))
             yield map
-
-
-class GoldenPathDataProvider(DictDataProvider):
-    '''
-    Class that defines a data provider for AGP datasets.
-    '''
-    settings: Dict[str, str] = {}
-
-    def __init__(self, source, **kwargs):
-        self.goldenpath = AGPFile(source)
-        self.column_names = [str(gp) for gp in self.goldenpath]
-        super().__init__(source, **kwargs)
-
-    def __iter__(self):
-        for gp in self.goldenpath:
-            yield gp
